Continue embedding the bluetooth protocol on both sides #4
it does not work at all

+++ ArduinoMazeSolver/Motor.h
@@ -2,7 +2,7 @@
 #include <Arduino.h>
 #include "Direction.h"
 
-void moveMotorOnSide(Direction side, Direction orientation, int speed)
+void moveMotorOnSide(Direction side, Direction orientation, byte speed)
 {
 	speed = max(min(speed, 255), 0);
 
@@ -10,7 +10,7 @@ void moveMotorOnSide(Direction side, Direction orientation, int speed)
 	analogWrite(side == left ? 3 : 11, speed);
 }
 
-void moveBothMotors(int speedLeft, Direction orientationLeft, int speedRight, Direction orientationRight)
+void moveBothMotors(byte speedLeft, Direction orientationLeft, byte speedRight, Direction orientationRight)
 {
 	moveMotorOnSide(left, orientationLeft, speedLeft);
 	moveMotorOnSide(right, orientationRight, speedRight);
